:recycle: [refactor]: refactor code in main page
