+++ src/components/commons/ModalStyle.js
@@ -1,0 +1,9 @@
+export const ModalStyle = {
+  overlay: {
+    display: "flex",
+    alignItems: "center",
+    justifyContent: "center",
+    backgroundColor: "rgba(0,0,0,0.65)",
+    zIndex: 10,
+  },
+};
