--- src/data.json
@@ -1,48 +1,0 @@
-{
-  "userId": 3,
-  "templateId": "52396ab8-0188-4000-b30d-0f448926c80f",
-  "title": "Changed title",
-  "subtitle": "sample subtitle2",
-  "isStar": true,
-  "groupId": 4,
-  "group": {
-    "userId": 3,
-    "id": 4,
-    "name": "회사",
-    "color": "#38D9A9"
-  },
-  "content": [
-    {
-      "id": "kxprp3s9xfqaj1dgsrd",
-      "tag": "p",
-      "flag": 1,
-      "html": "000님 안녕하세요, 저는 00팀 000입니다."
-    },
-    {
-      "id": "kxprufeacvsuthhot4",
-      "tag": "p",
-      "flag": 1,
-      "html": "00 프로젝트와 관련해서 00팀이 주최하는 회의를 진행하고자 합니다.<div>이번 회의일정은 다음과 같습니다.</div>"
-    },
-    {
-      "id": "kxprp3s9xfqaj1dgard",
-      "tag": "p",
-      "flag": 1,
-      "html": "1. 일시: 0000년 00월 00일 오전/오후 00시 ~  00시 <div>2. 장소: (장소)</div><div>3. 안건: (안건)</div>4. 참석 대상: (참석 대상)<div>5. 사전 준비사항: (사전 준비사항)</div>"
-    },
-    {
-      "id": "kxydhiwx4nuqllb33iw",
-      "tag": "p",
-      "flag": 1,
-      "html": "부득이하게 본 회의에 참석이 어려우신 경우, 원활한 회의 진행을 위해 00월 00일 오전/오후 00시까지 회신 부탁드립니다."
-    },
-    {
-      "id": "kxydi9ryaymjilivbli",
-      "tag": "p",
-      "flag": 1,
-      "html": "감사합니다.<div>000 드림.</div>"
-    }
-  ],
-  "createdAt": "2022-01-10T14:23:04.395897Z",
-  "updatedAt": "2022-01-10T14:41:09.901046Z"
-}
