--- src/fetchedData.json
@@ -1,17 +1,0 @@
-[
-  {
-    "title": "회사",
-    "color": "#3138FF",
-    "id": 1
-  },
-  {
-    "title": "학교",
-    "color": "#FA5252",
-    "id": 2
-  },
-  {
-    "title": "외주",
-    "color": "#38D9A9",
-    "id": 3
-  }
-]
